Move DatabaseEngineUniqueConstraintViolationException to exceptions
